flask-minimal/requirements.txt
Flask==3.1.0
requests==2.32.3
python-dotenv==1.0.1
flask-cors==5.0.0
gunicorn==23.0.0
Werkzeug==3.1.3
